admin panel

+++ recipe-app/delete_post.php
@@ -1,10 +1,13 @@
 <?php 
+session_start();
 include("db.php");
 
-if(isset($_GET['id'])){
+if(isset($_GET['id']) && isset($_SESSION['user_id'])){
     $post_id = $_GET['id'];
-    $run = $conn->prepare("delete from posts where post_id = :post");
+    $user_id = $_SESSION['user_id'];
+    $run = $conn->prepare("delete from posts where post_id = :post AND user_id =:user);
     $run->bindParam(':post', $post_id);
+    $run->bindParam(':user', $user_id);
     $result =$run->execute();
     if($result){
